reissue 토큰 검증 조건문 수정(and -> or)

+++ src/main/java/project/lincook/backend/service/AuthService.java
@@ -63,7 +63,7 @@ public class AuthService {
 		}
 
 		// 요청된 RT의 유효성 검사 & Redis에 저장되어 있는 RT와 같은지 비교
-		if(!jwtTokenProvider.validateRefreshToken(requestRefreshToken) || !refreshTokenInRedis.equals(requestRefreshToken)) {
+		if(!jwtTokenProvider.validateRefreshToken(requestRefreshToken) && !refreshTokenInRedis.equals(requestRefreshToken)) {
 			redisService.deleteValues("RT(" + SERVER + "):" + principal); // 탈취 가능성 -> 삭제
 			return null; // -> 재로그인 요청
 		}
